@ Phase 1: 拆 Core/Twse 兩專案 + 補 IStockMarketService 介面
結構（行為不變，邏輯重寫留待 Phase 2）：
- 新增 src/TWStockLib.Core（零 TWSE 依賴）：Models、Cache、Observer、
  Extension、IStockMarketFactory/IDataFetchStrategy 介面、StockMarketService
- 新增 src/TWStockLib.Twse（依賴 Core）：TwseMarketFactory、
  TwseDataFetchStrategy、ServiceCollectionExtensions
- 新增 IStockMarketService 介面，StockMarketService 實作之
  （修掉舊 README 提及卻不存在的介面）
- API 模型移至 Twse/Internal 並改 internal；刪除未使用的 TWSEAPIModel
- DI：AddStockServices → AddTwStock，舊名保留為 [Obsolete] 轉呼叫；
  改註冊 IStockMarketService
- 套件策略：Twse 專案 PackageId 維持 TWStockLib（現有使用者無痛升級，
  自動帶入 TWStockLib.Core）
- 更新 .slnx、publish.yml pack 路徑、TestExample 參考與服務解析

驗證：slnx build 兩框架 0 error；本機 pack 兩套件成功（TWStockLib 依賴
TWStockLib.Core 1.1.0）；TestExample 端到端跑通（2450 股、TSE/OTC 歷史、即時報價）

@
