+++ src/TWStockLib.Core/Cache/ICacheService.cs
