+++ src/TWStockLib.Core/Extension/EnumExtension.cs
