+++ src/TWStockLib.Core/IDataFetchStrategy.cs
