+++ src/TWStockLib.Core/IStockMarketFactory.cs
