+++ src/TWStockLib.Core/Models/MarketType.cs
