+++ src/TWStockLib.Core/Models/StockData.cs
