+++ src/TWStockLib.Core/Models/StockHistory.cs
