+++ src/TWStockLib.Core/Models/StockQuote.cs
