+++ src/TWStockLib.Core/Observer/DefaultProvidedObservers/ConsoleStockPriceObserver.cs
