+++ src/TWStockLib.Core/Observer/IStockPriceObserver.cs
