+++ src/TWStockLib.Core/TWStockLib.Core.csproj
@@ -1,0 +1,22 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>net8.0;net9.0</TargetFrameworks>
+    <RootNamespace>TWStockLib</RootNamespace>
+
+    <!-- 套件專屬資訊（共用部分見 ..\..\Directory.Build.props） -->
+    <IsPackable>true</IsPackable>
+    <PackageId>TWStockLib.Core</PackageId>
+    <Version>1.1.0</Version>
+    <Description>TWStockLib 核心抽象與模型（無 TWSE 來源依賴）。</Description>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Caching.Memory" Version="[8.0.1,)" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="[8.0.0,)" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="..\..\README.md" Pack="true" PackagePath="\" />
+  </ItemGroup>
+</Project>
