+++ src/TWStockLib.Twse/TwseMarketFactory.cs
